base algolia functions to algolia.js
